cli: Add BaseCLI base class with shared ExitCode and error handling
Introduce picard/cli/base.py with:

- ExitCode(IntEnum): Standard exit codes (SUCCESS=0, ERROR=1,
  NOT_FOUND=2, CANCELLED=130) shared across all subcommands.
- BaseCLI: Base class providing common infrastructure:
  - _is_debug_mode(): Check --debug flag
  - _handle_exception(e, message): Error output + traceback in debug mode
  - run(): Top-level wrapper with KeyboardInterrupt/Exception handling
  - _dispatch(): Abstract method for subclass command routing

PluginCLI now inherits from BaseCLI, removing ~20 lines of duplicated
boilerplate (ExitCode definition, _is_debug_mode, _handle_exception,
try/except in run). Its dispatch logic moves to _dispatch().
